Allow rules to be swapped in and out
This is a bit weird, but the idea behind this was to introduce the rules gradually,
so not all boards would have the full set of rules. For example, you could ignore
capturing to start with, then another example would show capturing other player's
stones, then another example prevents suicide while allowing snapbacks.

Each rule is basically a function that takes a bunch of state and maybe returns
a new state. I had to make this a type instead of a type alias because otherwise
the aliases for MoveInProgress and Game would be mutually recursive.

This introduces cycling between the two players and preventing a player from taking
two turns.