Very crappy capture function
This is incomplete at the moment

What should happen is that all the neighbouring enemy stones
are considered as groups, so you need to surround the whole thing
to capture them.

Also, we need to remove the annotations when the stone they're on dissapears.

I think this needs refactoring before that: Board is not a very useful type because
it doesn't encapsulate the size of the board, and that's really important for
determining how things connect.