Extract application types
Details about the AI in use in belong to the app, not the game itself.  Create
a parent model to contain everything. This doesn't have a good name yet, maybe
it should just be "GoApp" or something.

Make generateMove just return a command and export a function to lookup
strategies by name instead of exposing a union type. This is so that we can
just specify the ai with a flag. Remove the AI alias because it is useless
noise.

The AI module could probably be a more structured set of submodules like Game.
There is a lot of update logic in Main that could be extracted out. Hmmm.