Extract module for identifying groups and liberties
Move everything group-related out of the Board module into a new Group
module. This keeps the board module small and focused just on the low-level
placing and removing of stones and annotations.

The Group module is a layer on top that identifies groups of connected
stones. This can be used by the Game modules to remove groups that are considered dead - that is, those that have zero liberties.

Everything in the Board module now returns Lists for internal consistency.
This might change to Sets later on. The Group module uses Sets exclusively.
This is more convenient for the implementation and makes more sense for the
caller.